requirements.txt
Django==1.6
Markdown==2.3.1
South==0.8.3
amqp==1.3.3
anyjson==0.3.3
argparse==1.2.1
beautifulsoup4==4.3.2
billiard==3.3.0.7
celery==3.1.4
crontab==0.18
django-celery==3.1.1
django-common-helpers==0.6.0
django-cron==0.3.2
kombu==3.0.5
python-crontab==1.5.1
python-dateutil==2.2
pytz==2013.8
simplejson==3.3.1
six==1.4.1
wsgiref==0.1.2
